refactor: consolidate chart schemas and tools
