+++ src/schemas/base.ts
@@ -15,60 +15,6 @@ export const BaseConfigSchema = {
   axisYTitle: z.string().optional().describe("Set the y-axis title of chart."),
 };
 
-// Common data schema components
-export const LineChartDataSchema = z.object({
-  time: z.string(),
-  value: z.number(),
-});
-
-export const ColumnChartDataSchema = z.object({
-  category: z.string(),
-  value: z.number(),
-  group: z.string().optional(),
-});
-
-export const PieChartDataSchema = z.object({
-  category: z.string(),
-  value: z.number(),
-});
-
-export const AreaChartDataSchema = z.object({
-  time: z.string(),
-  value: z.number(),
-  group: z.string().optional(),
-});
-
-export const BarChartDataSchema = z.object({
-  category: z.string(),
-  value: z.number(),
-  group: z.string().optional(),
-});
-
-export const ScatterChartDataSchema = z.object({
-  x: z.number(),
-  y: z.number(),
-});
-
-export const WordCloudChartDataSchema = z.object({
-  text: z.string(),
-  value: z.string(),
-});
-
-export const RadarChartDataSchema = z.object({
-  name: z.string(),
-  value: z.number(),
-  group: z.string().optional(),
-});
-
-// Define recursive schema for hierarchical data
-export const TreeNodeSchema: z.ZodType<any> = z.lazy(() =>
-  z.object({
-    name: z.string(),
-    value: z.number(),
-    children: z.array(TreeNodeSchema).optional(),
-  }),
-);
-
 export const NodeSchema = z.object({
   name: z.string(),
 });
@@ -77,30 +23,4 @@ export const EdgeSchema = z.object({
   source: z.string(),
   target: z.string(),
   name: z.string(),
-});
-
-export const MindMapNodeSchema: z.ZodType<any> = z.lazy(() =>
-  z.object({
-    name: z.string(),
-    children: z.array(MindMapNodeSchema).optional(),
-  }),
-);
-
-export const FishboneNodeSchema: z.ZodType<any> = z.lazy(() =>
-  z.object({
-    name: z.string(),
-    children: z.array(FishboneNodeSchema).optional(),
-  }),
-);
-
-export const DualAxesSeriesSchema = z.object({
-  type: z
-    .enum(["column", "line"])
-    .describe("The optional value can be 'column' or 'line'."),
-  data: z
-    .array(z.number())
-    .describe(
-      "When type is column, the data represents quantities. When type is line, the data represents ratios.",
-    ),
-  axisYTitle: z.string().describe("Set the y-axis title of the chart series."),
 });
