+++ src/schemas/charts/area.ts
@@ -1,0 +1,36 @@
+import { z } from "zod";
+import { zodToJsonSchema } from "zod-to-json-schema";
+import { BaseConfigSchema } from "../base";
+
+// Area chart data schema
+const AreaChartDataSchema = z.object({
+  time: z.string(),
+  value: z.number(),
+  group: z.string().optional(),
+});
+
+// Area chart input schema
+export const AreaChartInputSchema = z.object({
+  data: z
+    .array(AreaChartDataSchema)
+    .describe("Data for area chart, such as, [{ time: '2018', value: 99.9 }]."),
+  stack: z
+    .boolean()
+    .optional()
+    .describe(
+      "Whether stacking is enabled. When enabled, area charts require a 'group' field in the data.",
+    ),
+  width: BaseConfigSchema.width.optional(),
+  height: BaseConfigSchema.height.optional(),
+  title: BaseConfigSchema.title,
+  axisXTitle: BaseConfigSchema.axisXTitle,
+  axisYTitle: BaseConfigSchema.axisYTitle,
+});
+
+// Area chart tool descriptor
+export const AreaChartTool = {
+  name: "generate_area_chart",
+  description:
+    "Generate a area chart to show data trends under continuous independent variables and observe the overall data trend, such as, displacement = velocity (average or instantaneous) × time: s = v × t. If the x-axis is time (t) and the y-axis is velocity (v) at each moment, an area chart allows you to observe the trend of velocity over time and infer the distance traveled by the area's size.",
+  inputSchema: zodToJsonSchema(AreaChartInputSchema),
+};
