+++ src/schemas/charts/histogram.ts
@@ -1,0 +1,29 @@
+import { z } from "zod";
+import { zodToJsonSchema } from "zod-to-json-schema";
+import { BaseConfigSchema } from "../base";
+
+// Histogram chart input schema
+export const HistogramChartInputSchema = z.object({
+  data: z
+    .array(z.number())
+    .describe("Data for histogram chart, such as, [78, 88, 60, 100, 95]."),
+  binNumber: z
+    .number()
+    .optional()
+    .describe(
+      "Number of intervals to define the number of intervals in a histogram.",
+    ),
+  width: BaseConfigSchema.width.optional(),
+  height: BaseConfigSchema.height.optional(),
+  title: BaseConfigSchema.title,
+  axisXTitle: BaseConfigSchema.axisXTitle,
+  axisYTitle: BaseConfigSchema.axisYTitle,
+});
+
+// Histogram chart tool descriptor
+export const HistogramChartTool = {
+  name: "generate_histogram_chart",
+  description:
+    "Generate a histogram chart to show the frequency of data points within a certain range. It can observe data distribution, such as, normal and skewed distributions, and identify data concentration areas and extreme points.",
+  inputSchema: zodToJsonSchema(HistogramChartInputSchema),
+};
