+++ src/schemas/index.ts
@@ -1,21 +1,38 @@
-// Combine all schemas for easy reference
-import * as ChartSchemas from "./charts";
+import {
+  LineChartInputSchema,
+  ColumnChartInputSchema,
+  PieChartInputSchema,
+  AreaChartInputSchema,
+  BarChartInputSchema,
+  HistogramChartInputSchema,
+  ScatterChartInputSchema,
+  WordCloudChartInputSchema,
+  RadarChartInputSchema,
+  TreemapChartInputSchema,
+  DualAxesChartInputSchema,
+  MindMapChartInputSchema,
+  NetworkGraphInputSchema,
+  FlowDiagramInputSchema,
+  FishboneDiagramInputSchema,
+} from "./charts";
 
-// Export the map of chart types to their respective schemas
+// Export the map of chart types to their respective schemas (for backward compatibility)
 export const SchemaMap = {
-  line: ChartSchemas.LineChartInputSchema,
-  column: ChartSchemas.ColumnChartInputSchema,
-  pie: ChartSchemas.PieChartInputSchema,
-  area: ChartSchemas.AreaChartInputSchema,
-  bar: ChartSchemas.BarChartInputSchema,
-  histogram: ChartSchemas.HistogramChartInputSchema,
-  scatter: ChartSchemas.ScatterChartInputSchema,
-  "word-cloud": ChartSchemas.WordCloudChartInputSchema,
-  radar: ChartSchemas.RadarChartInputSchema,
-  treemap: ChartSchemas.TreemapChartInputSchema,
-  "dual-axes": ChartSchemas.DualAxesChartInputSchema,
-  "mind-map": ChartSchemas.MindMapChartInputSchema,
-  "network-graph": ChartSchemas.NetworkGraphInputSchema,
-  "flow-diagram": ChartSchemas.FlowDiagramInputSchema,
-  "fishbone-diagram": ChartSchemas.FishboneDiagramInputSchema,
+  line: LineChartInputSchema,
+  column: ColumnChartInputSchema,
+  pie: PieChartInputSchema,
+  area: AreaChartInputSchema,
+  bar: BarChartInputSchema,
+  histogram: HistogramChartInputSchema,
+  scatter: ScatterChartInputSchema,
+  "word-cloud": WordCloudChartInputSchema,
+  radar: RadarChartInputSchema,
+  treemap: TreemapChartInputSchema,
+  "dual-axes": DualAxesChartInputSchema,
+  "mind-map": MindMapChartInputSchema,
+  "network-graph": NetworkGraphInputSchema,
+  "flow-diagram": FlowDiagramInputSchema,
+  "fishbone-diagram": FishboneDiagramInputSchema,
 };
+
+export { ChartTools } from "./charts";
