+++ src/services/stdio.ts
@@ -4,5 +4,5 @@ import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js"
 export async function startStdioMcpServer(server: Server): Promise<void> {
   const transport = new StdioServerTransport();
   await server.connect(transport);
-  console.debug("Code Runner MCP Server running on stdio");
+  console.debug("MCP Server Chart running on stdio");
 }
